requirement.txt
anyio==3.7.0
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
cryptography==41.0.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.3
django-allauth==0.57.0
django-import-export==3.3.3
djangorestframework==3.14.0
ecdsa==0.18.0
et-xmlfile==1.1.0
fastapi==0.95.2
greenlet==2.0.2
idna==3.4
MarkupPy==1.14
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.0.1
psycopg2==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.8
PyJWT==2.8.0
pypng==0.20220715.0
python-dateutil==1.5
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
sqlparse==0.4.4
starlette==0.27.0
stripe==7.0.0
tablib==3.5.0
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.0.3
xlrd==2.0.1
xlwt==1.3.0
